___JmolVersion="13.3.9_dev_2013.11.27d2"
code: JavaScript Integer, Long, Short, Byte, Float, Double all reworked 



Former-commit-id: 87f0622b0db74ae756892db4ab5bb7fce6075883

+++ Jmol/src/org/jmol/adapter/smarter/AtomSetCollectionReader.java
@@ -147,7 +147,7 @@ public abstract class AtomSetCollectionReader {
   // buffer
   public String line, prevline;
   protected int[] next = new int[1];
-  protected long ptLine;
+  protected int ptLine;
 
   // protected/public state variables
   public int[] latticeCells;
