+++ Jmol/src/org/jmol/util/Logger.java
@@ -298,22 +298,22 @@ public final class Logger {
     }
   }
 
-  static Map<String,Long>htTiming = new Hashtable<String, Long>();
+  static Map<String, Integer>htTiming = new Hashtable<String, Integer>();
   public static void startTimer(String msg) {
     if (msg != null)
-      htTiming.put(msg, Long.valueOf(System.currentTimeMillis()));
+      htTiming.put(msg, Integer.valueOf((int) System.currentTimeMillis()));
   }
 
-  public static String getTimerMsg(String msg, long time) {
+  public static String getTimerMsg(String msg, int time) {
     if (time == 0)
       time = getTimeFrom(msg);
     return "Time for " + msg + ": " + (time) + " ms";
   }
   
   private static int getTimeFrom(String msg) {
-    Long t;
+    Integer t;
     return (msg == null || (t = htTiming.get(msg)) == null ? -1 : (int) (System
-        .currentTimeMillis() - t.longValue()));
+        .currentTimeMillis() - t.intValue()));
   }
 
   public static int checkTimer(String msg, boolean andReset) {
@@ -326,16 +326,16 @@ public final class Logger {
   }
   
   public static void checkMemory() {
-    long bTotal = 0, bFree = 0, bMax = 0;
+    int bTotal = 0, bFree = 0, bMax = 0;
     /**
      * @j2sIgnore
      */
     {
       Runtime runtime = Runtime.getRuntime();
       runtime.gc();
-      bTotal = runtime.totalMemory();
-      bFree = runtime.freeMemory();
-      bMax = runtime.maxMemory();
+      bTotal = (int)runtime.totalMemory();
+      bFree = (int)runtime.freeMemory();
+      bMax = (int)runtime.maxMemory();
     }
     info("Memory: Total-Free="+ (bTotal - bFree)+"; Total=" +  bTotal + "; Free=" + bFree 
         + "; Max=" + bMax);
